+++ main.tf
@@ -14,10 +14,14 @@ provider "aws" {
 }
 
 
+/**
+* Note: This module takes 2 applies currently because of the
+* the creation and then data reference, this is only really a problem in this example
+*/
 module "role" {
-  source           = "./modules/base_role"
-  role_name        = "my-managed-cloud-role"
-  role_policy_name = "default"
+  source        = "./modules/base_policy_role"
+  new_role_name = "my-managed-cloud-role"
+  policy_name   = "default"
 }
 
 
@@ -29,7 +33,7 @@ module "tiered_storage" {
     Environment = "Prod"
   }
   # attach policy to existing role
-  existing_role_name = module.role.role_name
+  existing_role_name = module.role.role_names[0]
 }
 
 module "aws_vault" {
@@ -39,5 +43,5 @@ module "aws_vault" {
     Enviroment : "Production"
   }
   # attach policy to existing role
-  existing_role_name = module.role.role_name
+  existing_role_name = module.role.role_names[0]
 }